Added dedicated file for cw content page skeleton
