+++ packages/commonwealth/client/scripts/views/components/component_kit/CWContentPage/index.tsx
@@ -1,0 +1,1 @@
+export * from './CWContentPage';
