+++ packages/commonwealth/client/scripts/views/pages/view_thread/ViewThreadPage.tsx
@@ -34,8 +34,8 @@ import Permissions from '../../../utils/Permissions';
 import { CreateComment } from '../../components/Comments/CreateComment';
 import { Select } from '../../components/Select';
 import { CWCheckbox } from '../../components/component_kit/cw_checkbox';
-import type { SidebarComponents } from '../../components/component_kit/cw_content_page';
-import { CWContentPage } from '../../components/component_kit/cw_content_page';
+import type { SidebarComponents } from '../../components/component_kit/CWContentPage';
+import { CWContentPage } from '../../components/component_kit/CWContentPage';
 import { CWIcon } from '../../components/component_kit/cw_icons/cw_icon';
 import { CWText } from '../../components/component_kit/cw_text';
 import { CWTextInput } from '../../components/component_kit/cw_text_input';
