+++ packages/commonwealth/client/scripts/views/pages/view_thread/linked_proposals_card.tsx
@@ -15,7 +15,7 @@ import type ChainEntity from '../../../models/ChainEntity';
 import type Thread from '../../../models/Thread';
 import type { ThreadStage } from '../../../models/types';
 import { CWButton } from '../../components/component_kit/cw_button';
-import { CWContentPageCard } from '../../components/component_kit/cw_content_page';
+import { CWContentPageCard } from '../../components/component_kit/CWContentPage';
 import { Modal } from '../../components/component_kit/cw_modal';
 import { CWSpinner } from '../../components/component_kit/cw_spinner';
 import { CWText } from '../../components/component_kit/cw_text';
