+++ packages/commonwealth/client/scripts/views/pages/view_thread/linked_threads_card.tsx
@@ -10,7 +10,7 @@ import { slugify } from '../../../../../shared/utils';
 import type Thread from '../../../models/Thread';
 import { LinkSource } from '../../../models/Thread';
 import { CWButton } from '../../components/component_kit/cw_button';
-import { CWContentPageCard } from '../../components/component_kit/cw_content_page';
+import { CWContentPageCard } from '../../components/component_kit/CWContentPage';
 import { Modal } from '../../components/component_kit/cw_modal';
 import { CWText } from '../../components/component_kit/cw_text';
 import { LinkedThreadModal } from '../../modals/linked_thread_modal';
