+++ packages/commonwealth/client/styles/components/component_kit/CWContentPage.scss
